sync: verified self-review command + self-found fixes + drop free tier
- self-review: reviewer → adversarial verification pipeline with json_object
  finalizer, budget partitioning, scope discipline, and a JSONL run journal
- fixes surfaced by self-review: telemetry avg-tokens float division,
  orchestrator UTF-8-safe summary truncation, acceptance timeout classification
- router: drop the weak :free tier from the ladder

+++ internal/telemetry/hints.go
@@ -52,6 +52,6 @@ func Summary(records []Record) string {
 		cost += r.Cost
 	}
 	n := len(records)
-	return fmt.Sprintf("last %d runs: %d%% ok · avg %.1f steps · avg %d tok · ~$%.4f total",
-		n, ok*100/n, float64(steps)/float64(n), tokens/n, cost)
+	return fmt.Sprintf("last %d runs: %d%% ok · avg %.1f steps · avg %.1f tok · ~$%.4f total",
+		n, ok*100/n, float64(steps)/float64(n), float64(tokens)/float64(n), cost)
 }
